(refactor) move SidebarTopOverlay & SidebarLeftOverlay to Navigation
