+++ packages/Frontend/src/components/Navigation/index.js
@@ -1,0 +1,4 @@
+import SidebarTopOverlay from './SidebarTopOverlay';
+import SidebarLeftOverlay from './SidebarLeftOverlay';
+
+export { SidebarLeftOverlay, SidebarTopOverlay };
